Move the semver package to common and alter references.
