+++ tests/unit/test_config.py
@@ -21,7 +21,7 @@ from dbt.adapters.postgres import PostgresCredentials
 from dbt.adapters.contracts.connection import QueryComment, DEFAULT_QUERY_COMMENT
 from dbt.contracts.project import PackageConfig, LocalPackage, GitPackage
 from dbt.node_types import NodeType
-from dbt.semver import VersionSpecifier
+from dbt.common.semver import VersionSpecifier
 from dbt.task.base import ConfiguredTask
 
 from dbt.flags import set_from_args
