+++ tests/unit/test_semver.py
@@ -3,7 +3,7 @@ import itertools
 
 from typing import List
 from dbt.common.exceptions import VersionsNotCompatibleError
-from dbt.semver import (
+from dbt.common.semver import (
     VersionSpecifier,
     UnboundedVersionSpecifier,
     VersionRange,
